Prom_shim/requirements.txt
blinker==1.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
prometheus-flask-exporter==0.23.0
prometheus_client==0.20.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
